fix: correct depth/encode/metadata bugs
- depth: fix OpenVINO folder resolution (double-dash name)
- depth: add missing return in OGDepthV3Cuda near-empty-mask branch (frame desync)
- ffmpeg: guard color-space metadata read when file absent
- metadata: assign ColorSpace/PixelFormat from correct ffprobe fields
- rife ncnn: correct inverted half/dtype selection
- ytdlp: fix typoed title info-dict key

+++ src/unifiedInterpolate.py
@@ -1190,7 +1190,7 @@ class RifeNCNN:
                 .reshape(self.height, self.width, 3)
                 .permute(2, 0, 1)
                 .unsqueeze(0)
-                .to(dtype=torch.float32 if self.half else torch.float16)
+                .to(dtype=torch.float16 if self.half else torch.float32)
                 .mul(1 / 255.0)
             )
             interpQueue.put(output)
